+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## NAACCR XML Version History
 
-**Version 10.3 (not released yet)**
+**Version 11.0 (not released yet)**
 
-- Added support for new dateTime type.
+- Added support for NAACCR 25.
+- Added support for new specification version 1.7 (added support for new dateTime type).
 
 **Version 10.2**
 
@@ -18,7 +19,7 @@
 **Version 10.0**
 
 - Added support for NAACCR 24.
-- Added support for new specification version 1.7 (removed trim attribute and some padding options); only NAACCR 24 dictionary will use that version for now.
+- Added support for new specification version 1.7 (removed trim attribute and some padding options); only NAACCR 24 dictionary will use that version for now).
 - Removed standalone program (the removal was announced a year ago when support for NAACCR 23 was added).
 - Re-wrote the SAS macro to use a temp fixed-column file instead of a temp CSV file to resolve a maxed-out line length issue.